refactor(gpu): move ToCudaDType into its own header, out of cuFill_gpu.hpp
ToCudaDType is a generic scalar-to-device-type trait with no
connection to Fill specifically; cuTrace_internal.cu including
cuFill_gpu.hpp only for this trait meant the trace implementation
depended on an unrelated operation's header.

Moves the trait to a new cuTypeTraits_gpu.hpp, leaving cuFill_gpu.hpp
with just the FillGpu declaration it's named for. cuFill_gpu.cu and
cuTrace_internal.cu now include the trait header directly instead of
picking it up as a side effect of including cuFill_gpu.hpp.
