+++ src/backend/linalg_internal_gpu/cuTrace_internal.cu
@@ -2,9 +2,10 @@
 #include "cytnx_error.hpp"
 #include "Tensor.hpp"
 #include "backend/Storage.hpp"
-#include "backend/utils_internal_gpu/cuFill_gpu.hpp"
+#include "backend/utils_internal_gpu/cuTypeTraits_gpu.hpp"
 
 #include <algorithm>
+#include <iterator>
 #include <limits>
 #include <vector>
