+++ src/backend/utils_internal_gpu/cuFill_gpu.cu
@@ -1,4 +1,5 @@
 #include "backend/utils_internal_gpu/cuFill_gpu.hpp"
+#include "backend/utils_internal_gpu/cuTypeTraits_gpu.hpp"
 
 #include <complex>
