+++ src/backend/utils_internal_gpu/cuFill_gpu.hpp
@@ -3,40 +3,8 @@
 
 #include "Type.hpp"
 
-#ifdef UNI_GPU
-  #include <complex>
-
-  #include "cuda/std/complex"
-#endif
-
 namespace cytnx {
   namespace utils_internal {
-
-#ifdef UNI_GPU
-    /// @brief Maps a cytnx/std scalar type to the device arithmetic type used
-    /// inside CUDA kernels. Complex types map to the bit-compatible
-    /// cuda::std::complex, which provides device operator+ /
-    /// construction-from-zero; every other type passes through unchanged.
-    template <typename DType>
-    struct ToCudaDType {
-      typedef DType type;
-    };
-
-    template <typename DType>
-    struct ToCudaDType<std::complex<DType>> {
-      typedef cuda::std::complex<DType> type;
-    };
-
-    template <>
-    struct ToCudaDType<cytnx_complex128> {
-      typedef cuda::std::complex<double> type;
-    };
-
-    template <>
-    struct ToCudaDType<cytnx_complex64> {
-      typedef cuda::std::complex<float> type;
-    };
-#endif
 
     /**
      * @brief Assign the given value to the first `count` elements in the range beginning at
